Add files via upload
this is the initial code for vacuum of new version of pcb without modbus and units conversion for torr and pascal if vacuum values are beynd the limit of 4 digit. if the vacuum values were exceeds more than 4 digit in integer or decimal. it will display on exponential form. vacuum calibration done through ui  only by pressing sensor's min and max buttons.